Improve protected routing
